Fix some warnings that SQLAlchemy 1.0 was printing
It was complaining about the following scenario: Evaluation has two
columns (submission_id, dataset_id), with three foreign key constraints
built on top of them and a relationship associated to each of them (to
SubmissionResult but also, individually, to Submission and Dataset). It
was warning us that by changing the submission or dataset relationship
we would write on one of the columns and have the unwanted side-effect
of also changing the submission_result. To fix it we make the submission
and dataset relationship become "view only": they can be used as a
shortcut to directly access the object but they cannot be used to change
it; the only way to change them is to change the submission_result.