Restyle language selector to browser default
Changed bootstrap.css so that it doesn't specify a default border style on
<select> elements. On most browsers, specifying any border style whatsoever
will result in it not rendering the dropdown as default, hence, it is
impossible to specify styling elements on individual elements to prevent this
behaviour.

Bootstrap is only used on CWS, and within it, <select> is only used in one
other place; namely the i18n language selector box, which overrides this style
anyway. It doesn't seem super-critical and any future <select> elements that
desire this border style (1px solid #BBB) should manually apply this style.

Tested on Chrome and FF and renders fine on both.