Display output of first compiler command only
Currently, CMS displays the stdout/stderr of the last compilation command that
is run. We are electing to instead display the stdout/stderr of the first
such command, as the first command is usually the most vital.

Currently, only for Python do we use multiple commands of which the second (and
last) is a /usr/bin/mv operation. In the case when the Python compilation
fails, the output file will not be produced so the /usr/bin/mv will also fail.
As this second failure is likely to confuse inexperienced users, we elect to
only show the output of the first command, and not all of them.